Refactor frontend directory structure
