+++ frontend/src/components/search/SearchResult.js
