+++ frontend/src/components/search/search.css
