+++ frontend/src/components/session/LogInForm.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import { Link, withRouter } from 'react-router-dom'
 import { connect } from 'react-redux'
 import { Button, Container, Form } from 'react-bootstrap'
-import { login } from '../actions/sessions'
+import { login } from '../../actions/sessions'
 
 
 class LogInForm extends Component {
